refactor(backend): relocate Markdown tests
